Move QPI::collection tests to separate file
